node.js 8/10 for travis
more refactoring

add package-lock.json

fix package-lock.json?

it was the cyclical dependency
