+++ .travis.yml
@@ -3,8 +3,8 @@ sudo: false
 notifications:
   email: false
 node_js:
-  - "6"
   - "8"
+  - "10"
 cache:
   directories: # Cache dependencies
     - node_modules
